requirements.txt
streamlit>=1.28.0
langchain>=0.0.350
langchain-community>=0.0.10
ollama>=0.1.7
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
pypdf>=3.17.0
python-docx>=0.8.11
openpyxl>=3.1.0
streamlit-chat>=0.1.1
watchdog>=3.0.0
requests>=2.31.0